move filters to their respective containers
